Shoot direction implemented

+++ game_state.h
@@ -8,5 +8,5 @@ struct game_state
   bool started;
   d2d_object player;
   std::list<std::unique_ptr<d2d_object>> bullets;
-  int targetPosX, targetPosY;
+  double targetPosX, targetPosY;
 };
